Allow user-defined group identifier
